requirements.txt
numpy
torch
torchvision
scikit-learn
matplotlib
pillow